requirements.txt
config==0.5.1
matplotlib==3.7.1
numpy==1.23.5
numpy==1.24.4
pandas==1.5.3
pandas==2.0.2
plotly==5.15.0
python-dotenv==1.0.1
scipy==1.13.1
snowflake_connector_python==3.7.1
snowflake_snowpark_python==1.14.0
streamlit==1.33.0
scipy
snowflake==0.7.0
openpyxl
xlrd
st_aggrid